fix(speaker-detail): move action sheets to speaker detail
