+++ src/components/page-speaker-detail/page-speaker-detail.tsx
@@ -1,3 +1,4 @@
+import { actionSheetController, getMode } from '@ionic/core';
 import { Component, Prop, h } from '@stencil/core';
 import { ConferenceData } from '../../providers/conference-data';
 import { Plugins } from '@capacitor/core';
@@ -24,6 +25,65 @@ export class PageSpeakerDetail {
     });
   }
 
+  async openSpeakerShare(speaker: any) {
+    const actionSheet = await actionSheetController.create({
+      header: 'Share ' + speaker.name,
+      buttons: [
+        {
+          text: 'Copy Link',
+          handler: () => {
+            console.log(
+              'Copy link clicked on https://twitter.com/' + speaker.twitter
+            );
+            if (
+              (window as any)['cordova'] &&
+              (window as any)['cordova'].plugins.clipboard
+            ) {
+              (window as any)['cordova'].plugins.clipboard.copy(
+                'https://twitter.com/' + speaker.twitter
+              );
+            }
+          }
+        },
+        {
+          text: 'Share via ...'
+        },
+        {
+          text: 'Cancel',
+          role: 'cancel'
+        }
+      ]
+    });
+
+    actionSheet.present();
+  }
+
+  async openContact(speaker: any) {
+    const mode = getMode();
+
+    const actionSheet = await actionSheetController.create({
+      header: 'Contact ' + speaker.name,
+      buttons: [
+        {
+          text: `Email ( ${speaker.email} )`,
+          icon: mode !== 'ios' ? 'mail' : null,
+          handler: () => {
+            window.open('mailto:' + speaker.email);
+          }
+        },
+        {
+          text: `Call ( ${speaker.phone} )`,
+          icon: mode !== 'ios' ? 'call' : null,
+          handler: () => {
+            window.open('tel:' + speaker.phone);
+          }
+        }
+      ]
+    });
+
+    actionSheet.present();
+  }
+
   render() {
     const speaker = this.speaker;
 
@@ -35,7 +95,12 @@ export class PageSpeakerDetail {
               <ion-back-button defaultHref="/speakers"></ion-back-button>
             </ion-buttons>
             <ion-buttons slot="end">
-
+            <ion-button onClick={() => this.openContact(speaker)}>
+              <ion-icon slot="icon-only" ios="call-outline" md="call-sharp"></ion-icon>
+            </ion-button>
+            <ion-button onClick={() => this.openSpeakerShare(speaker)}>
+              <ion-icon slot="icon-only" ios="share-outline" md="share-sharp"></ion-icon>
+            </ion-button>
             </ion-buttons>
           </ion-toolbar>
         </ion-header>
